provide default view for all tool calls
