+++ src/inspect_ai/approval/_apply.py
@@ -1,8 +1,11 @@
+import pprint
 from contextvars import ContextVar
+from textwrap import indent
+from typing import Any
 
 from inspect_ai.approval._approval import Approval
 from inspect_ai.model._call_tools import ToolDef
-from inspect_ai.tool._tool_call import ToolCall
+from inspect_ai.tool._tool_call import ToolCall, ToolCallContent, ToolCallView
 
 from ._approver import Approver
 from ._policy import ApprovalPolicy, policy_approver
@@ -15,13 +18,28 @@ async def apply_tool_approval(
 
     approver = _tool_approver.get(None)
     if approver:
+        # create default view
+        view = ToolCallView(
+            call=ToolCallContent(
+                format="markdown",
+                content="```python\n"
+                + format_function_call(tool_call.function, tool_call.arguments)
+                + "\n```\n",
+            )
+        )
+
+        # current sample state
         state = sample_state()
+
+        # call approver
         approval = await approver(
             content=content,
             call=tool_call,
-            view=None,
+            view=view,
             state=state,
         )
+
+        # process decision
         match approval.decision:
             case "approve":
                 return True, approval
@@ -31,8 +49,35 @@ async def apply_tool_approval(
                 if state:
                     state.completed = True
                 return False, approval
+
+    # no approval system registered
     else:
         return True, None
+
+
+def format_function_call(
+    func_name: str, args_dict: dict[str, Any], indent_spaces: int = 4, width: int = 80
+) -> str:
+    formatted_args = []
+    for key, value in args_dict.items():
+        formatted_value = format_value(value, width)
+        formatted_args.append(f"{key}={formatted_value}")
+
+    args_str = ", ".join(formatted_args)
+
+    if len(args_str) <= width - 1 - len(func_name) - 2:  # 2 for parentheses
+        return f"{func_name}({args_str})"
+    else:
+        indented_args = indent(",\n".join(formatted_args), " " * indent_spaces)
+        return f"{func_name}(\n{indented_args}\n)"
+
+
+def format_value(value: object, width: int) -> str:
+    if isinstance(value, str):
+        return f"'{value}'"
+    elif isinstance(value, list | tuple | dict):
+        return pprint.pformat(value, width=width)
+    return str(value)
 
 
 def init_tool_approval(approval: list[ApprovalPolicy] | None) -> None:
