+++ src/inspect_ai/approval/_approver.py
@@ -13,7 +13,7 @@ class Approver(Protocol):
         self,
         content: str,
         call: ToolCall,
-        view: ToolCallView | None = None,
+        view: ToolCallView,
         state: TaskState | None = None,
     ) -> Approval:
         """
